feat(workspace-cap): extract WorkspaceSearchConfig data struct (slice 8 of N preparatory)
ADR-152 section 3 F3.4 preparatory slice. Verbatim port of the
WorkspaceSearchConfig data struct (fields + Default impl) from
desktop-client/ironclaw/src/config/search.rs into a new config module in
dasclaw_workspace_cap. Environment-variable resolution stays in ironclaw
config layer (it depends on host-specific helpers and ConfigError); the
inherent impl is converted to a free function resolve() per Rust's orphan
rule. The Workspace::with_search_config parameter type is updated to refer
to the cap crate path directly, removing the last direct dependency
between Workspace and crate::config.

Closes #706
Refs #674
